:boom: breaking: change VideoScene API to improve performance

+++ src/generic_scene.rs
@@ -18,9 +18,9 @@ enum SceneEnum {
 
 #[wasm_bindgen]
 pub struct GenericScene {
-    width: u64,
-    height: u64,
-    fps: u64,
+    width: u32,
+    height: u32,
+    fps: u32,
     scene: SceneEnum
 }
 
@@ -109,15 +109,15 @@ impl GenericScene {
         }
     }
     #[wasm_bindgen(js_name = getFps)]
-    pub fn get_fps(&self) -> u64 {
+    pub fn get_fps(&self) -> u32 {
         return self.fps;
     }
     #[wasm_bindgen(js_name = getHeight)]
-    pub fn get_height(&self) -> u64 {
+    pub fn get_height(&self) -> u32 {
         return self.height;
     }
     #[wasm_bindgen(js_name = getWidth)]
-    pub fn get_width(&self) -> u64 {
+    pub fn get_width(&self) -> u32 {
         return self.width;
     }
     #[wasm_bindgen(js_name = renderFrame)]
@@ -432,7 +432,7 @@ impl GenericScene {
         &mut self,
         animation_func: Function,
         object_indices: Vec<usize>,
-        duration_in_frames: u64,
+        duration_in_frames: u32,
         rate_func: Function
     ) {
         match &mut self.scene {
@@ -473,7 +473,7 @@ impl GenericScene {
         }
     }
     #[wasm_bindgen(js_name = wait)]
-    pub async fn wait(&mut self, duration_in_frames: u64) {
+    pub async fn wait(&mut self, duration_in_frames: u32) {
         match &mut self.scene {
             #[cfg(feature = "browser")]
             SceneEnum::Scene(scene) => {
